SM:MM: Add support for zcopy transfers
zcopy for SM is implemented by remapping physical pages from sender
process to receiver other. In order for this functionality to be supported,
the shuffle.ko module must be loaded into the kernel.

In this implementation the pages are exchanged between two processes.
Meaning, the receiver process must provide allocated pages to exchange with
the sender.

This version supports only one page being exchanged.
The hellow_world example returns INPROGRESS and the remapped memory is
released later.

Signed-off-by: Tanya Brokhman <tanya.brokhman@huawei.com>